backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pyjwt==2.8.0
cryptography==41.0.7
python-jose==3.3.0
Werkzeug==3.0.1